# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Airport Flight Data Collector API
  description: |
    Comprehensive REST and GraphQL API for accessing flight data, predictions, and analytics.
    Features include real-time flight status, ML-based delay predictions, pattern detection,
    data export, webhook integrations, and advanced analytics for SFO, YYZ, and YVR airports.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/czhaoca/airport-flight-data-collector/issues
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.flightdata.example.com/v1
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /flights:
    get:
      tags:
        - Flights
      summary: Get flight data
      description: Retrieve flight data with various filters
      parameters:
        - $ref: '#/components/parameters/Airport'
        - $ref: '#/components/parameters/FlightType'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Airline'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightListResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
...
  /export/flights:
    get:
      tags:
        - Export
      summary: Export flight data
      description: Export flight data in various formats
      parameters:
        - $ref: '#/components/parameters/Airport'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - name: format
          in: query
          description: Export format
          schema:
            type: string
            enum: [json, csv, parquet]
            default: json
        - name: fields
          in: query
          description: Comma-separated list of fields to include
          schema:
            type: string
            example: 'flightNumber,airline,status,delay'
      responses:
        '200':
          description: Flight data export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
            text/csv:
              schema:
                type: string
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /export/statistics:
    get:
      tags:
        - Export
      summary: Export statistics
      description: Export statistical summaries and analytics
      parameters:
        - $ref: '#/components/parameters/Period'
        - $ref: '#/components/parameters/Airport'
        - name: format
          in: query
          description: Export format
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Statistics export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsExportResponse'
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Airport:
      name: airport
      in: query
      description: Filter by airport code
      schema:
        type: string
        enum: [SFO, YYZ, YVR]
    FlightType:
      name: type
      in: query
      description: Filter by flight type
      schema:
        type: string
        enum: [arrival, departure]
    StartDate:
      name: startDate
      in: query
      description: Start date (inclusive)
      schema:
        type: string
        format: date
        example: '2025-01-01'
    EndDate:
      name: endDate
      in: query
      description: End date (inclusive)
      schema:
        type: string
        format: date
        example: '2025-01-31'
    Status:
      name: status
      in: query
      description: Filter by flight status
      schema:
        type: string
        enum: [ON_TIME, DELAYED, CANCELLED, DEPARTED, ARRIVED, BOARDING]
    Airline:
      name: airline
      in: query
      description: Filter by airline code
      schema:
        type: string
        example: 'UA'
    Limit:
      name: limit
      in: query
      description: Maximum number of results
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    Sort:
      name: sort
      in: query
      description: Sort order
      schema:
        type: string
        enum: [scheduledTime, actualTime, flightNumber, status]
        default: scheduledTime
    Period:
      name: period
      in: query
      description: Time period for statistics
      schema:
        type: string
        enum: [24h, 7d, 30d, 90d, 1y]
        default: 7d
  schemas:
...
    Flight:
      type: object
      properties:
        id:
          type: string
          description: Unique flight identifier
        flightNumber:
          type: string
          example: 'UA123'
        airline:
          $ref: '#/components/schemas/Airline'
        origin:
          type: string
          example: 'LAX'
        destination:
          type: string
          example: 'SFO'
        scheduledTime:
          type: string
          format: date-time
        actualTime:
          type: string
          format: date-time
          nullable: true
        estimatedTime:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [ON_TIME, DELAYED, CANCELLED, DEPARTED, ARRIVED, BOARDING]
        gate:
          type: string
          nullable: true
        terminal:
          type: string
          nullable: true
        type:
          type: string
          enum: [arrival, departure]
        aircraft:
          type: string
          nullable: true
        delay:
          type: integer
          description: Delay in minutes
          nullable: true
    Airline:
      type: object
      properties:
        code:
          type: string
          example: 'UA'
        name:
          type: string
          example: 'United Airlines'
...
    FlightListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Flight'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        cached:
          type: boolean
          default: false
...
    ExportResponse:
      type: object
      properties:
        success:
          type: boolean
        exportId:
          type: string
        downloadUrl:
          type: string
        format:
          type: string
        recordCount:
          type: integer
        fileSize:
          type: integer
          description: File size in bytes
        expiresAt:
          type: string
          format: date-time
    StatisticsExportResponse:
      type: object
      properties:
        success:
          type: boolean
        exportId:
          type: string
        downloadUrl:
          type: string
        format:
          type: string
        categories:
          type: array
          items:
            type: string
        period:
          type: string
        expiresAt:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: BAD_REQUEST
            message: Invalid date format
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: UNAUTHORIZED
            message: Invalid or missing authentication token
...
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when rate limit resets
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: RATE_LIMIT_EXCEEDED
            message: Too many requests. Please try again later.